requirements.txt
langchain==0.1.0
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
google-search-results==2.4.2
python-weather==1.1.1
